+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export { delayExecution, type CancelToken } from './delayExecution';
 export { throttle } from './throttle';
 export { isClient } from './isClient';
 export { isTouchDevice } from './isTouchDevice';
+export { validators } from './validators';
